requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
anthropic>=0.28.0
python-bring-api>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
